[image-manipulator] Export missing types, remove fallback, register web module
