+++ packages/expo-image-manipulator/src/ImageManipulator.types.ts
@@ -1,4 +1,7 @@
-import type { NativeModule, SharedObject, SharedRef } from 'expo';
+import type { NativeModule } from 'expo';
+
+import type { ImageManipulatorContext } from './ImageManipulatorContext';
+import ImageRef from './ImageRef';
 
 // @needsAudit
 export type ImageResult = {
@@ -122,90 +125,18 @@ export type SaveOptions = {
 };
 
 export declare class ImageManipulator extends NativeModule {
-  Context: typeof Context;
+  /**
+   * @hidden
+   */
+  Context: typeof ImageManipulatorContext;
+
+  /**
+   * @hidden
+   */
   Image: typeof ImageRef;
 
   /**
    * Loads an image from the given URI and creates a new image manipulation context.
    */
-  manipulate(uri: string): Context;
-
-  /**
-   * @hidden This was removed from the native code on iOS and Web. See `ImageManipulator.ts`
-   * for the JS implementation that uses the new API under the hood.
-   * @platform android
-   */
-  manipulateAsync(uri: string, actions: Action[], saveOptions: SaveOptions);
-}
-
-/**
- * A context for an image manipulation. It provides synchronous, chainable functions that schedule transformations on the original image to the background thread.
- * Use an asynchronous [`renderAsync`](#contextrenderasync) to await for all transformations to finish and access the final image.
- */
-export declare class Context extends SharedObject {
-  /**
-   * Resizes the image to the given size.
-   * @param size Values correspond to the result image dimensions. If you specify only one value, the other will
-   * be calculated automatically to preserve image ratio.
-   */
-  resize(size: { width?: number | null; height?: number | null }): Context;
-
-  /**
-   * Rotates the image by the given number of degrees.
-   * @param degrees Degrees to rotate the image. Rotation is clockwise when the value is positive and
-   * counter-clockwise when negative.
-   */
-  rotate(degrees: number): Context;
-
-  /**
-   * Flips the image vertically or horizontally.
-   * @param flipType An axis on which image will be flipped. Only one flip per transformation is available. If you
-   * want to flip according to both axes then provide two separate transformations.
-   */
-  flip(flipType: 'vertical' | 'horizontal'): Context;
-
-  /**
-   * Crops the image to the given rectangle's origin and size.
-   * @param rect Fields specify top-left corner and dimensions of a crop rectangle.
-   */
-  crop(rect: ActionCrop['crop']): Context;
-
-  /**
-   * Set the image size and offset. If the image is enlarged, unfilled areas are set to the `backgroundColor`.
-   * To position the image, use `originX` and `originY`.
-   *
-   * @platform web
-   */
-  extent(options: ActionExtent['extent']): Context;
-
-  /**
-   * Resets the manipulator context to the originally loaded image.
-   */
-  reset(): Context;
-
-  /**
-   * Awaits for all manipulation tasks to finish and resolves with a reference to the resulted native image.
-   */
-  renderAsync(): Promise<ImageRef>;
-}
-
-/**
- * A reference to a native instance of the image.
- */
-export declare class ImageRef extends SharedRef<'image'> {
-  /**
-   * Width of the image.
-   */
-  width: number;
-
-  /**
-   * Height of the image.
-   */
-  height: number;
-
-  /**
-   * Saves the image to the file system in the cache directory.
-   * @param options A map defining how modified image should be saved.
-   */
-  saveAsync(options?: SaveOptions): Promise<ImageResult>;
+  manipulate(uri: string): ImageManipulatorContext;
 }
