+++ packages/expo-image-manipulator/src/NativeImageManipulatorModule.ts
@@ -1,4 +1,4 @@
-import { requireNativeModule } from 'expo-modules-core';
+import { requireNativeModule } from 'expo';
 
 import { ImageManipulator } from './ImageManipulator.types';
